Created 403 page, updated ToDoController.java

+++ src/main/java/com/softserve/itacademy/controller/ToDoController.java
@@ -3,6 +3,7 @@ package com.softserve.itacademy.controller;
 import com.softserve.itacademy.model.Task;
 import com.softserve.itacademy.model.ToDo;
 import com.softserve.itacademy.model.User;
+import com.softserve.itacademy.repository.UserRepository;
 import com.softserve.itacademy.service.TaskService;
 import com.softserve.itacademy.service.ToDoService;
 import com.softserve.itacademy.service.UserService;
@@ -12,6 +13,7 @@ import org.springframework.validation.BindingResult;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
+import java.security.Principal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -23,18 +25,24 @@ public class ToDoController {
     private final ToDoService todoService;
     private final TaskService taskService;
     private final UserService userService;
+    private final UserRepository userRepository;
 
-    public ToDoController(ToDoService todoService, TaskService taskService, UserService userService) {
+    public ToDoController(ToDoService todoService, TaskService taskService, UserService userService, UserRepository userRepository) {
         this.todoService = todoService;
         this.taskService = taskService;
         this.userService = userService;
+        this.userRepository = userRepository;
     }
 
     @GetMapping("/create/users/{owner_id}")
-    public String create(@PathVariable("owner_id") long ownerId, Model model) {
-        model.addAttribute("todo", new ToDo());
-        model.addAttribute("ownerId", ownerId);
-        return "create-todo";
+    public String create(@PathVariable("owner_id") long ownerId, Model model, Principal principal) {
+        User user = userRepository.findByEmail(principal.getName());
+        if (userService.readById(ownerId).getEmail().equals(principal.getName()) || user.getRole().getName().equals("ADMIN")) {
+            model.addAttribute("todo", new ToDo());
+            model.addAttribute("ownerId", ownerId);
+            return "create-todo";
+        }
+        return "redirect:/access-denied";
     }
 
     @PostMapping("/create/users/{owner_id}")
@@ -49,22 +57,30 @@ public class ToDoController {
     }
 
     @GetMapping("/{id}/tasks")
-    public String read(@PathVariable long id, Model model) {
-        ToDo todo = todoService.readById(id);
-        List<Task> tasks = taskService.getByTodoId(id);
-        List<User> users = userService.getAll().stream()
-                .filter(user -> user.getId() != todo.getOwner().getId()).collect(Collectors.toList());
-        model.addAttribute("todo", todo);
-        model.addAttribute("tasks", tasks);
-        model.addAttribute("users", users);
-        return "todo-tasks";
+    public String read(@PathVariable long id, Model model, Principal principal) {
+        User user = userRepository.findByEmail(principal.getName());
+        if (todoService.readById(id).getOwner().getEmail().equals(principal.getName()) || user.getRole().getName().equals("ADMIN")) {
+            ToDo todo = todoService.readById(id);
+            List<Task> tasks = taskService.getByTodoId(id);
+            List<User> users = userService.getAll().stream()
+                    .filter(u -> u.getId() != todo.getOwner().getId()).collect(Collectors.toList());
+            model.addAttribute("todo", todo);
+            model.addAttribute("tasks", tasks);
+            model.addAttribute("users", users);
+            return "todo-tasks";
+        }
+        return "redirect:/access-denied";
     }
 
     @GetMapping("/{todo_id}/update/users/{owner_id}")
-    public String update(@PathVariable("todo_id") long todoId, @PathVariable("owner_id") long ownerId, Model model) {
-        ToDo todo = todoService.readById(todoId);
-        model.addAttribute("todo", todo);
-        return "update-todo";
+    public String update(@PathVariable("todo_id") long todoId, @PathVariable("owner_id") long ownerId, Model model, Principal principal) {
+        User user = userRepository.findByEmail(principal.getName());
+        if (userService.readById(ownerId).getEmail().equals(principal.getName()) || user.getRole().getName().equals("ADMIN")) {
+            ToDo todo = todoService.readById(todoId);
+            model.addAttribute("todo", todo);
+            return "update-todo";
+        }
+        return "redirect:/access-denied";
     }
 
     @PostMapping("/{todo_id}/update/users/{owner_id}")
@@ -82,17 +98,25 @@ public class ToDoController {
     }
 
     @GetMapping("/{todo_id}/delete/users/{owner_id}")
-    public String delete(@PathVariable("todo_id") long todoId, @PathVariable("owner_id") long ownerId) {
-        todoService.delete(todoId);
-        return "redirect:/todos/all/users/" + ownerId;
+    public String delete(@PathVariable("todo_id") long todoId, @PathVariable("owner_id") long ownerId, Principal principal) {
+        User user = userRepository.findByEmail(principal.getName());
+        if (userService.readById(ownerId).getEmail().equals(principal.getName()) || user.getRole().getName().equals("ADMIN")) {
+            todoService.delete(todoId);
+            return "redirect:/todos/all/users/" + ownerId;
+        }
+        return "redirect:/access-denied";
     }
 
     @GetMapping("/all/users/{user_id}")
-    public String getAll(@PathVariable("user_id") long userId, Model model) {
-        List<ToDo> todos = todoService.getByUserId(userId);
-        model.addAttribute("todos", todos);
-        model.addAttribute("user", userService.readById(userId));
-        return "todos-user";
+    public String getAll(@PathVariable("user_id") long userId, Model model, Principal principal) {
+        User user = userRepository.findByEmail(principal.getName());
+        if (userService.readById(userId).getEmail().equals(principal.getName()) || user.getRole().getName().equals("ADMIN")) {
+            List<ToDo> todos = todoService.getByUserId(userId);
+            model.addAttribute("todos", todos);
+            model.addAttribute("user", userService.readById(userId));
+            return "todos-user";
+        }
+        return "redirect:/access-denied";
     }
 
     @GetMapping("/{id}/add")
